organization 7

+++ RiskLabAI/features/feature_importance/feature_importance_mda.py
@@ -60,7 +60,7 @@ class FeatureImportanceMDA(FeatureImportanceStrategy):
 
         cv_generator = KFold(n_splits=self.n_splits)
         baseline_scores = pd.Series(dtype=float)
-        shuffled_scores = pd.DataFrame(columns=x.columns)
+        shuffled_scores = pd.DataFrame(columns=x.columns, dtype=float)
 
         for i, (train_idx, test_idx) in enumerate(cv_generator.split(x)):
             print(f"Fold {i} start ...")
